% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\docType{data}
\name{contacts}
\alias{contacts}
\title{Data set 'contacts'}
\format{
A JSON string with ragged, nested contact details
}
\usage{
contacts
}
\description{
Data set 'contacts'
}
\keyword{datasets}
