% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{db_create}
\alias{db_create}
\title{Read documents.}
\usage{
db_create(conn, doc)
}
\description{
Read documents.
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
db_create(conn, doc)
}
}

