% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{docdb_update.src_sqlite}
\alias{docdb_update.src_sqlite}
\title{#' @export
docdb_update.src_mongo <- function(src, key, value, ...) {

assert(value, 'data.frame')

if (!key %in% attr(src, "dbs")) sofa::db_create(src[\link{1}], dbname = key)
sofa::db_bulk_update(src[\link{1}], dbname = key, doc = value, ...)

}}
\usage{
\method{docdb_update}{src_sqlite}(src, key, value, ...)
}
\description{
#' @export
docdb_update.src_mongo <- function(src, key, value, ...) {

assert(value, 'data.frame')

if (!key %in% attr(src, "dbs")) sofa::db_create(src[\link{1}], dbname = key)
sofa::db_bulk_update(src[\link{1}], dbname = key, doc = value, ...)

}
}
