% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docdbi-package.R
\docType{package}
\name{docdbi-package}
\alias{docdbi}
\alias{docdbi-package}
\title{Document database connector}
\description{
Document database connector
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Rich FitzJohn \email{rich.fitzjohn@gmail.com}

Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}
\keyword{package}

