% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docdbi-package.R
\docType{package}
\name{docplyr-package}
\alias{docplyr}
\alias{docplyr-package}
\title{Query, manipulate, etc. data in document DB's}
\description{
Currently supported DB's are:
}
\details{
\itemize{
 \item Elasticsearch
 \item Solr
 \item CouchDB
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

