% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.R
\name{get}
\alias{get}
\title{Read documents.}
\usage{
get(conn, docid)
}
\description{
Read documents.
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
res <- create(conn, doc)
get(conn, res$id)
}
}

