% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\docType{data}
\name{mapdata}
\alias{mapdata}
\title{mapdata JSON data set}
\format{
A JSON string with ragged, nested travel details
}
\usage{
mapdata
}
\description{
mapdata JSON data set
}
\keyword{datasets}
