% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\name{nodbi-defunct}
\alias{nodbi-defunct}
\title{Defunct functions in nodbi}
\description{
\itemize{
\item \link{src_etcd}: etcd removed, with all its S3 methods for \verb{docdb_*}
\item \link{src_redis}: redis removed, with all its S3 methods for \verb{docdb_*}
}
}
