% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src}
\alias{src}
\title{Setup database connections}
\description{
Setup database connections
}
\details{
There is a `src_*()` function to setup a connection to each
of the database backends. Each has their own unique set of parameters.

- MongoDB - [src_mongo()]
- CouchDB - [src_couchdb()]
- Etcd - [src_etcd()]
- Elasticsearch - [src_elastic()]
}
