% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src_elasticsearch.R
\name{src_elasticsearch}
\alias{src_elasticsearch}
\title{Setup an Elasticsearch database connection}
\usage{
src_elasticsearch(base = "http://127.0.0.1", port = 9200, user = NULL,
  pwd = NULL, key = NULL, force = FALSE, ...)
}
\arguments{
\item{base}{the base url, defaults to localhost (http://127.0.0.1)}

\item{port}{port to connect to, defaults to 9200 (optional)}

\item{user}{User name, if required for the connection. You can specify, but ignored for now.}

\item{pwd}{Password, if required for the connection. You can specify, but ignored for now.}

\item{key}{An API key, ignored for now}

\item{force}{Force re-load of connection details}

\item{...}{Further args passed on to print for the es_conn class.}
}
\description{
Setup an Elasticsearch database connection
}
\examples{
\dontrun{
src_elasticsearch()
}
}

