% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_etcd.R
\name{src_etcd}
\alias{src_etcd}
\title{Setup an Etcd database connection}
\usage{
src_etcd(host = "127.0.0.1", port = 2379, api_version = "v2",
  allow_redirect = TRUE, scheme = "http")
}
\description{
Setup an Etcd database connection
}
\examples{
\dontrun{
(src <- src_etcd())
class(src)
}
}
