% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_mongo.R
\name{src_mongo}
\alias{src_mongo}
\title{Setup a mongoDB database connection}
\usage{
src_mongo(collection = "test", db = "test", url = "mongodb://localhost",
  ...)
}
\arguments{
\item{collection}{(character) name of collection}

\item{db}{(character) name of database}

\item{url}{(character) address of the mongodb server in mongo connection
string URI format.}

\item{verbose}{(logical) emit some more output. default: \code{TRUE}}

\item{options}{(list) additional connection options such as SSL keys/certs}
}
\description{
Setup a mongoDB database connection
}
\examples{
\dontrun{
(con <- src_mongo())
print(con)
}
}
