% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_riak.R
\name{src_riak}
\alias{src_riak}
\title{Setup a Riak database connection}
\usage{
src_riak(bucket = "test", host = "127.0.0.1", port = 8098, path = NULL,
  transport = "http", user = NULL, pwd = NULL, headers = NULL)
}
\arguments{
\item{host}{(character) host value, default: 127.0.0.1}

\item{port}{(integer/numeric) Port. Remember that if you don't want a port
set, set this parameter to NULL. Default: 8098}

\item{path}{(character) context path that is appended to the end of the url. e.g.,
bar in http://foo.com/bar. Default: NULL, ignored}

\item{transport}{(character) http or https. Default: http}

\item{user}{(character) Username, if any}

\item{pwd}{(character) Password, if any}

\item{headers}{(list) list of named headers}
}
\description{
Setup a Riak database connection
}
\details{
Riak's "databases" are called buckets - and a bucket holds
keys. If there is no bucket named "test" the default bucket name,
we'll create it for you in this function call by inserting a key
with value of \code{""} into a bucket called "test". There's no way to
create a bucket other than inserting a key, so there you have it. You can
of course do this yourself.
}
\examples{
\dontrun{
x <- src_riak()
}
}
