% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_redis.R
\name{src_rlite}
\alias{src_rlite}
\title{Setup an rlite database connection}
\usage{
src_rlite(filename = ":memory:")
}
\arguments{
\item{filename}{Name of the database to use}
}
\description{
Setup an rlite database connection
}
\examples{
con <- src_rlite()
class(con)
}

