% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src_rrlite.R
\name{src_rrlite}
\alias{src_rrlite}
\title{Setup an rrlite database connection}
\usage{
src_rrlite(filename = ":memory:", db = NULL)
}
\arguments{
\item{filename}{Name of the database to use}

\item{db}{Prefix for the keys}
}
\description{
Setup an rrlite database connection
}
\examples{
con <- src_rrlite()
class(con)
}

