% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\docType{data}
\name{contacts}
\alias{contacts}
\title{contacts JSON data set}
\format{
A JSON string with ragged, nested contact details
}
\usage{
contacts
}
\description{
contacts JSON data set
}
\keyword{data}
