% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{docdb_create}
\alias{docdb_create}
\title{Create a document}
\usage{
docdb_create(conn, docid)
}
\arguments{
\item{conn}{Connection object, result of call to src}

\item{docid}{Document ID}
}
\description{
Create a document
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
docdb_create(conn, doc)

doc2 <- fromJSON("http://api.gbif.org/v1/species/2704174")
docdb_create(conn, doc2)
}
}

