% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{docdb_create}
\alias{docdb_create}
\title{Create documents}
\usage{
docdb_create(src, key, value, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{A key. See Details.}

\item{value}{A value}

\item{...}{Ignored}
}
\description{
Create documents
}
\details{
Note that with etcd, you have to prefix a key with a forward slash.
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
docdb_create(conn, doc)

doc2 <- fromJSON("http://api.gbif.org/v1/species/2704174")
docdb_create(conn, doc2)

key="mtcars2"
value=mtcars
docdb_create(src, key, value)
docdb_get(src, key)

# etcd
src <- src_etcd()
docdb_create(src, "/hello", "world")
## a data.frame
docdb_create(src, key = "/newmtcars7", value = mtcars)
docdb_get(src, "/newmtcars7")

# Elasticsearch
src <- src_elasticsearch()
docdb_create(src, key = "mtcars", value = mtcars)
docdb_create(src, key = "iris", value = iris)
docdb_create(src, key = "diamonds_small", value = diamonds[1:3000L,])
}
}

