% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{docdb_create}
\alias{docdb_create}
\title{Create documents}
\usage{
docdb_create(src, key, value, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{A key}

\item{value}{A value}

\item{...}{Ignored for now}
}
\description{
Create documents
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
docdb_create(conn, doc)

doc2 <- fromJSON("http://api.gbif.org/v1/species/2704174")
docdb_create(conn, doc2)

key="mtcars2"
value=mtcars
docdb_create(src, key, value)
docdb_get(src, key)

# etcd
src <- src_etcd()
docdb_create(src, "hello", "world")
}
}

