% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{docdb_create}
\alias{docdb_create}
\title{Create documents}
\usage{
docdb_create(src, key, value, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{A key. See Details.}

\item{value}{A value}

\item{...}{Ignored}
}
\description{
Create documents
}
\details{
Note that with etcd, you have to prefix a key with a forward slash.
}
\examples{
\dontrun{
# CouchDB
src <- src_couchdb()
docdb_create(src, key="mtcars2", value=mtcars)
docdb_get(src, "mtcars2")

# Etcd
# src <- src_etcd()
# docdb_create(src, key = "/newmtcars7", value = mtcars)
# docdb_get(src, "/newmtcars7")

# Elasticsearch
src <- src_elasticsearch()
docdb_create(src, key = "mtcars", value = mtcars)
docdb_create(src, key = "iris", value = iris)
docdb_create(src, key = "diamonds_small", value = diamonds[1:3000L,])

# Redis
### server
src1 <- src_redis()
docdb_create(src1, key = "mtcars", value = mtcars)
docdb_get(src1, "mtcars")
docdb_delete(src1, "mtcars")

### serverless
src2 <- src_rlite()
docdb_create(src2, key = "mtcars", value = mtcars)
docdb_get(src2, "mtcars")
docdb_delete(src2, "mtcars")

# MongoDB
src <- src_mongo()
docdb_create(src, key = "mtcars", value = mtcars)
docdb_get(src, "mtcars")

# Riak
src <- src_riak()
docdb_create(src, key = "mtcars2", value = mtcars)
docdb_get(src, "mtcars2")
identical(mtcars, docdb_get(src, "mtcars2"))
}
}
