% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delete.R
\name{docdb_delete}
\alias{docdb_delete}
\title{Delete documents}
\usage{
docdb_delete(src, key, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{A key}

\item{...}{Ignored for now}
}
\description{
Delete documents
}
\examples{
\dontrun{
# couchdb
src <- src_couchdb()
docdb_create(src, "mtcars", mtcars)
docdb_get(src, "mtcars")
docdb_delete(src, "mtcars")

# etcd
src <- src_etcd()
docdb_create(src, "/iris", iris)
docdb_get(src, "/iris")
docdb_delete(src, "/iris")
}
}

