% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{docdb_delete}
\alias{docdb_delete}
\title{Delete documents}
\usage{
docdb_delete(src, key, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{A key}

\item{...}{Ignored for now}
}
\description{
Delete documents
}
\examples{
\dontrun{
# couchdb
(src <- src_couchdb())
docdb_create(src, "mtcars", mtcars)
docdb_get(src, "mtcars")
docdb_delete(src, "mtcars")

# etcd
# src <- src_etcd()
# docdb_create(src, "/iris", iris)
# docdb_get(src, "/iris")
# docdb_delete(src, "/iris")

# elasticsearch
src <- src_elasticsearch()
docdb_create(src, "iris", iris)
docdb_get(src, "iris")
docdb_delete(src, "iris")

# Redis
### server
src1 <- src_redis()
docdb_create(src1, key = "mtcars", value = mtcars)
docdb_get(src1, "mtcars")
docdb_delete(src1, "mtcars")

### serverless
src2 <- src_rlite()
docdb_create(src2, key = "mtcars", value = mtcars)
docdb_get(src2, "mtcars")
docdb_delete(src2, "mtcars")

# mongo
src <- src_mongo("stuff")
docdb_create(src, "iris", iris)
docdb_get(src, "iris")
docdb_delete(src)

# riak
src <- src_riak()
docdb_create(src, "iris", iris)
docdb_get(src, "iris")
docdb_delete(src, "iris")
}
}
