% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{docdb_exists}
\alias{docdb_exists}
\title{Check if a database exists}
\usage{
docdb_exists(src, key, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{(character) A key, ignored for mongo}

\item{...}{Ignored for now}
}
\value{
logical, \code{TRUE} or \code{FALSE}
}
\description{
Check if a database exists
}
\details{
Note that with etcd, you have to prefix a key with a
forward slash.
}
