% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.R
\name{docdb_get}
\alias{docdb_get}
\title{Get documents}
\usage{
docdb_get(conn, docid)
}
\description{
Get documents
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
res <- docdb_create(conn, doc)
docdb_get(conn, res$id)
}
}

