% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.R
\name{docdb_get}
\alias{docdb_get}
\title{Get documents}
\usage{
docdb_get(src, docid, ...)
}
\arguments{
\item{docid}{Document ID}

\item{conn}{Connection object, result of call to src}
}
\description{
Get documents
}
\examples{
\dontrun{
conn <- src_couchdb()
library("jsonlite")
doc <- fromJSON("http://api.gbif.org/v1/species/2704179")
res <- docdb_create(conn, doc)
docdb_get(conn, res$id)

doc2 <- fromJSON("http://api.gbif.org/v1/species/2704174")
docout <- docdb_create(conn, doc2)
docdb_get(conn, docout$id)

# entire data.frame
conn <- src_couchdb()
docout <- docdb_create(conn, mtcars)
docdb_get(conn, pluck(docout, "id", "")[1])
docdb_get(conn, pluck(docout, "id", "")[1:5])
docdb_get(conn, pluck(docout, "id", ""))
}
}

