% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{docdb_get}
\alias{docdb_get}
\title{Get documents}
\usage{
docdb_get(src, docid, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{docid}{Document ID}

\item{...}{Ignored for now}
}
\description{
Get documents
}
\examples{
\dontrun{
# CouchDB
src <- src_couchdb()
docout <- docdb_create(src, key = "mtcars", value = mtcars)
docdb_get(src, "mtcars")

# Etcd
src <- src_etcd()
docdb_create(src, "/hello", "world")
docdb_get(src, "/hello")

# Elasticsearch
src <- src_elasticsearch()
docdb_create(src, "iris", iris)
docdb_get(src, "iris")

# Redis
src <- src_redis()
docdb_create(src, "mtcars", mtcars)
docdb_get(src, "mtcars")

# Mongo
src <- src_mongo()
docdb_create(src, "mtcars", mtcars)
docdb_get(src, "mtcars")
}
}
