% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{docdb_get}
\alias{docdb_get}
\title{Get documents}
\usage{
docdb_get(src, key, limit = NULL, ...)
}
\arguments{
\item{src}{source object, result of call to src}

\item{key}{(chartacter) A key. ignored for mongo}

\item{limit}{(integer) number of records/rows to return. by default
not passed, so you get all results. Only works for CouchDB,
Elasticsearch and MongoDB; ignored for others}

\item{...}{passed on to functions:
\itemize{
\item CouchDB: passed to \code{\link[sofa:db_alldocs]{sofa::db_alldocs()}}
\item etcd: passed to the \code{$key()} method
\item Elasticsearch: passed to \code{\link[elastic:Search]{elastic::Search()}}
\item Redis: ignored
\item MongoDB: ignored
}}
}
\description{
Get documents
}
\details{
Note that with etcd, you have to prefix a key with a
forward slash.
}
\examples{
\dontrun{
# CouchDB
src <- src_couchdb()
docout <- docdb_create(src, key = "mtcars2", value = mtcars)
docdb_get(src, "mtcars2")
docdb_get(src, "mtcars2", limit = 5)

# etcd
# src <- src_etcd()
# docdb_create(src, "/hello", mtcars)
# docdb_get(src, "/hello")

# Elasticsearch
src <- src_elastic()
docdb_create(src, "iris", iris)
docdb_get(src, "iris")
docdb_create(src, "d2", diamonds)
docdb_get(src, "d2", limit = 1010)

# Redis
src <- src_redis()
docdb_create(src, "mtcars", mtcars)
docdb_get(src, "mtcars")

# Mongo
src <- src_mongo()
docdb_create(src, "mtcars", mtcars)
docdb_get(src, "mtcars")
docdb_get(src, "mtcars", limit = 4)
}
}
