% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\docType{package}
\name{nodbi-package}
\alias{nodbi-package}
\alias{nodbi}
\title{Document database connector}
\description{
Supports NoSQL databases (Elasticsearch, CouchDB, MongoDB),
key-value databases (Redis), and other NoSQL types (etcd).
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}

Rich FitzJohn \email{rich.fitzjohn@gmail.com}

Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}
\keyword{package}
