% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\docType{package}
\name{nodbi-package}
\alias{nodbi-package}
\alias{nodbi}
\title{Document database connector}
\description{
Simplified document database access and manipulation,
providing a common API across supported 'NoSQL' databases
'Elasticsearch', 'CouchDB', 'MongoDB' as well as
'SQLite/JSON1' and 'PostgreSQL'.
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}

Rich FitzJohn \email{rich.fitzjohn@gmail.com}

Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}

Ralf Herold \email{ralf.herold@mailbox.org}
}
\keyword{package}
