% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodbi-package.R
\docType{package}
\name{nodbi-package}
\alias{nodbi}
\alias{nodbi-package}
\title{Document database connector}
\description{
Simplified document database access and manipulation,
providing a common API across supported 'NoSQL' databases
'Elasticsearch', 'CouchDB', 'MongoDB' as well as
'SQLite/JSON1', 'PostgreSQL' and 'DuckDB'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/nodbi/}
  \item \url{https://github.com/ropensci/nodbi}
  \item Report bugs at \url{https://github.com/ropensci/nodbi/issues}
}

}
\author{
Ralf Herold \email{ralf.herold@mailbox.org}

Scott Chamberlain \email{sckott@protonmail.com}

Rich FitzJohn \email{rich.fitzjohn@gmail.com}

Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}
\keyword{internal}
