% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src}
\alias{src}
\title{Setup database connections}
\description{
Setup database connections
}
\details{
There is a \verb{src_*()} function to setup a connection to each
of the database backends. Each has their own unique set of parameters.
\itemize{
\item MongoDB - \code{\link[=src_mongo]{src_mongo()}}
\item SQLite - \code{\link[=src_sqlite]{src_sqlite()}}
\item Elasticsearch - \code{\link[=src_elastic]{src_elastic()}}
\item CouchDB - \code{\link[=src_couchdb]{src_couchdb()}}
}

Documentation details for each database:
\itemize{
\item MongoDB - \url{https://docs.mongodb.com/}
\item SQLite/JSON1 - \url{https://www.sqlite.org/json1.html}
\item CouchDB - \url{http://docs.couchdb.org/}
\item Elasticsearch -
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/index.html}
}

Documentation for R packages used by \code{nodbi} for the databases:
\itemize{
\item mongolite - \url{https://CRAN.R-project.org/package=mongolite}
\item RSQLite - \url{https://CRAN.R-project.org/package=RSQLite}
\item sofa - \url{https://CRAN.R-project.org/package=sofa}
\item elastic - \url{https://CRAN.R-project.org/package=elastic}
}
}
