% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src}
\alias{src}
\title{Setup database connections}
\description{
Setup database connections
}
\details{
There is a \code{src_*()} function to setup a connection to each
of the database backends. Each has their own unique set of parameters.
\itemize{
\item MongoDB - \code{\link[=src_mongo]{src_mongo()}}
\item CouchDB - \code{\link[=src_couchdb]{src_couchdb()}}
\item Elasticsearch - \code{\link[=src_elastic]{src_elastic()}}
\item Redis - \code{\link[=src_redis]{src_redis()}}
\item SQLite - \code{\link[=src_sqlite]{src_sqlite()}}
}

Documentation details for each database:
\itemize{
\item MongoDB - \url{https://docs.mongodb.com/}
\item CouchDB - \url{http://docs.couchdb.org/}
\item Elasticsearch -
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current}
\item Redis - \url{https://redis.io/documentation}
\item SQLite/json1 - \url{https://www.sqlite.org/json1.html}
}
}
