% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\docType{data}
\name{src}
\alias{src}
\title{Setup database connections}
\description{
There is a \verb{src_*()} function to setup a connection to each
of the database backends. The backends may have specific parameters
in the respective function \verb{src_*()}, but all other \code{nodbi} functions
are independent of the backend (e.g., see \code{\link[=docdb_query]{docdb_query()}}).
}
\details{
\itemize{
\item MongoDB - \code{\link[=src_mongo]{src_mongo()}}
\item SQLite - \code{\link[=src_sqlite]{src_sqlite()}}
\item Elasticsearch - \code{\link[=src_elastic]{src_elastic()}}
\item CouchDB - \code{\link[=src_couchdb]{src_couchdb()}}
\item PostgreSQL - \code{\link[=src_postgres]{src_postgres()}}
\item DuckDB - \code{\link[=src_duckdb]{src_duckdb()}}
}

Documentation details for each database:
\itemize{
\item MongoDB - \url{https://docs.mongodb.com/}
\item SQLite/JSON1 - \url{https://www.sqlite.org/json1.html}
\item Elasticsearch -
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/index.html}
\item CouchDB - \url{http://docs.couchdb.org/}
\item PostgreSQL - \url{https://www.postgresql.org/docs/current/functions-json.html}
\item DuckDB - \url{https://duckdb.org/docs/extensions/json}
}

Documentation of R packages used by \code{nodbi} for the databases:
\itemize{
\item mongolite - \url{https://CRAN.R-project.org/package=mongolite}
\item RSQLite - \url{https://CRAN.R-project.org/package=RSQLite}
\item elastic - \url{https://CRAN.R-project.org/package=elastic}
\item sofa - \url{https://CRAN.R-project.org/package=sofa}
\item RPostgres - \url{https://CRAN.R-project.org/package=RPostgres}
\item duckdb - \url{https://CRAN.R-project.org/package=duckdb}
}
}
