% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src_couchdb.R
\name{src_couchdb}
\alias{src_couchdb}
\title{Setup a database connection}
\usage{
src_couchdb(type = "localhost", port = 5984, user = NULL, pwd = NULL)
}
\arguments{
\item{type}{One of localhost, cloudant, or iriscouch. This is what's used to determine
how to structure the URL to make the request.}

\item{port}{Port number}

\item{user}{Username, if any}

\item{pwd}{Password, if any}
}
\description{
Setup a database connection
}
\examples{
\dontrun{
src_couchdb()
}
}

