% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_couchdb.R
\name{src_couchdb}
\alias{src_couchdb}
\title{Setup a CouchDB database connection}
\usage{
src_couchdb(host = "127.0.0.1", port = 5984, path = NULL,
  transport = "http", user = NULL, pwd = NULL, headers = NULL)
}
\arguments{
\item{host}{(character) host value, default: 127.0.0.1}

\item{port}{(integer/numeric) Port. Remember that if you don't want a port
set, set this parameter to NULL. Default: 5984}

\item{path}{(character) context path that is appended to the end
of the url. e.g., bar in http://foo.com/bar. Default: NULL, ignored}

\item{transport}{(character) http or https. Default: http}

\item{user}{(character) Username, if any}

\item{pwd}{(character) Password, if any}

\item{headers}{(list) list of named headers}
}
\description{
Setup a CouchDB database connection
}
\details{
uses \pkg{sofa} under the hood; uses \link[sofa:Cushion]{sofa::Cushion} for
connecting
}
\examples{
\dontrun{
src_couchdb()
}
}
