% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_duckdb.R
\name{src_duckdb}
\alias{src_duckdb}
\title{Setup a DuckDB database connection}
\usage{
src_duckdb(drv = duckdb::duckdb(), dbdir = attr(drv, "dbdir"), ...)
}
\arguments{
\item{drv}{Object returned by \code{duckdb()}}

\item{dbdir}{Location for database files. Should be a path to an existing
directory in the file system. With the default, all
data is kept in RAM}

\item{...}{Additional named parameters passed on to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\description{
Setup a DuckDB database connection
}
\details{
Uses \code{\link[duckdb:duckdb]{duckdb::duckdb()}} under the hood
}
\examples{
\dontrun{
con <- src_duckdb()
print(con)
}
}
