% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_etcd.R
\name{src_etcd}
\alias{src_etcd}
\title{Setup an Etcd database connection}
\usage{
src_etcd()
}
\description{
Setup an Etcd database connection
}
\examples{
\dontrun{
(src <- src_etcd())
class(src)
}
}

