% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_mariadb.R
\name{src_mariadb}
\alias{src_mariadb}
\title{Setup a RMariaDB database connection}
\usage{
src_mariadb(dbname = "test", ...)
}
\arguments{
\item{dbname}{(character) name of database file,
defaults to ":memory:" for an in-memory database,
see \code{\link[RMariaDB:MariaDB]{RMariaDB::MariaDB()}}}

\item{...}{additional named parameters passed
on to \code{\link[RMariaDB:MariaDB]{RMariaDB::MariaDB()}}}
}
\value{
A \code{nodbi} source object
}
\description{
Setup a RMariaDB database connection
}
\details{
Uses \pkg{RSQLite} as backend. \pkg{nodbi} creates or uses
an SQLite table, with columns \verb{_id} and \code{json} created and used by
package \code{nodbi}, applying SQL functions as per
\url{https://www.sqlite.org/json1.html} to the \code{json} column.
Each row in the table represents a \code{JSON} document.
Any root-level \verb{_id} is extracted from the document(s) and used for
column \verb{_id}, otherwise a UUID is created as \verb{_id}.
The table is indexed on \verb{_id}.
For a benchmark, see \url{https://github.com/ropensci/nodbi#benchmark}
}
\examples{
\dontrun{
con <- src_sqlite()
print(con)
}

}
