% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_mongo.R
\name{src_mongo}
\alias{src_mongo}
\title{Setup a mongoDB database connection}
\usage{
src_mongo(url = "mongodb://localhost", db = "test")
}
\arguments{
\item{url}{Server host url}

\item{db}{Name of the database to use}
}
\description{
Setup a mongoDB database connection
}
\examples{
con <- src_mongo()
print(con)
}
