% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_mongo.R
\name{src_mongo}
\alias{src_mongo}
\title{Setup a MongoDB database connection}
\usage{
src_mongo(collection = "test", db = "test", url = "mongodb://localhost", ...)
}
\arguments{
\item{collection}{(character) Name of collection}

\item{db}{(character) Name of database}

\item{url}{(character) Address of the MongoDB server in Mongo connection
string URI format, see to \code{\link[mongolite:mongo]{mongolite::mongo()}}}

\item{...}{Additional named parameters passed on to \code{\link[mongolite:mongo]{mongolite::mongo()}}}
}
\description{
Setup a MongoDB database connection
}
\details{
Uses \pkg{monoglite} under the hood; uses \code{\link[mongolite:mongo]{mongolite::mongo()}} for
connecting
}
\examples{
\dontrun{
con <- src_mongo()
print(con)
}
}
