% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_mongo.R
\name{src_mongo}
\alias{src_mongo}
\title{Setup a MongoDB database connection}
\usage{
src_mongo(collection = "test", db = "test", url = "mongodb://localhost", ...)
}
\arguments{
\item{collection}{(character) name of collection}

\item{db}{(character) name of database}

\item{url}{(character) address of the MongoDB server in mongo connection
string URI format, see to \link[mongolite:mongo]{mongolite::mongo}}

\item{...}{additional named params passed on to \link[mongolite:mongo]{mongolite::mongo}}
}
\description{
Setup a MongoDB database connection
}
\details{
uses \pkg{monoglite} under the hood; uses \link[mongolite:mongo]{mongolite::mongo} for
connecting
}
\examples{
\dontrun{
con <- src_mongo()
print(con)
}
}
