% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_redis.R
\name{src_redis}
\alias{src_redis}
\title{Setup an Redis database connection}
\usage{
src_redis(host = "127.0.0.1", port = 6379)
}
\arguments{
\item{host}{Host URI}

\item{port}{Port number}
}
\description{
Setup an Redis database connection
}
\examples{
con <- src_redis()
class(con)
}

