% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_redis.R
\name{src_redis}
\alias{src_redis}
\title{Setup an Redis database connection}
\usage{
src_redis(...)
}
\arguments{
\item{...}{Redis connection configuration options. See \link[redux:redis_config]{redux::redis_config}
for more information}
}
\description{
Setup an Redis database connection
}
\details{
uses \pkg{redux} under the hood; uses \code{\link[redux:hiredis]{redux::hiredis()}} for
connecting
}
\examples{
\dontrun{
(con <- src_redis())
class(con)
}
}
