% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_redis.R
\name{src_rlite}
\alias{src_rlite}
\title{Setup an rlite database connection}
\usage{
src_rlite(...)
}
\arguments{
\item{...}{named config options passed on to
\code{\link[redux]{redis_config}}}
}
\description{
Setup an rlite database connection
}
\examples{
con <- src_rlite()
class(con)
}
